+++ robotbase/src/main/deploy/pathplanner/node8 to stage3.path
