+++ robotbase/src/main/deploy/pathplanner/stage3 to node6.path
