+++ robotbase/src/main/java/com/team2357/frc2023/trajectoryutil/AvailableTrajectoryCommands.java
@@ -1,11 +1,11 @@
 package com.team2357.frc2023.trajectoryutil;
 
-import com.team2357.frc2023.commands.auto.blue.grid1.BlueGridOneTwoConeAutoCommand;
-import com.team2357.frc2023.commands.auto.blue.grid1.BlueGridOneTwoConeBalanceAutoCommand;
-import com.team2357.frc2023.commands.auto.blue.grid2.BlueGridTwoTwoConeAutoCommand;
-import com.team2357.frc2023.commands.auto.blue.grid2.BlueGridTwoTwoConeBalanceAutoCommand;
-import com.team2357.frc2023.commands.auto.blue.grid3.BlueGridThreeTwoConeAutoCommand;
-import com.team2357.frc2023.commands.auto.blue.grid3.BlueGridThreeTwoConeBalanceAutoCommand;
+import com.team2357.frc2023.commands.auto.blue.gridone.BlueGridTwoTwoConeAutoCommand;
+import com.team2357.frc2023.commands.auto.blue.gridone.BlueGridTwoTwoConeBalanceAutoCommand;
+import com.team2357.frc2023.commands.auto.blue.gridtwo.BlueGridThreeTwoConeAutoCommand;
+import com.team2357.frc2023.commands.auto.blue.gridtwo.BlueGridThreeTwoConeBalanceAutoCommand;
+import com.team2357.frc2023.commands.auto.blue.gridzero.BlueGridOneTwoConeAutoCommand;
+import com.team2357.frc2023.commands.auto.blue.gridzero.BlueGridOneTwoConeBalanceAutoCommand;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
